+++ refinery-core/src/main/java/org/wikimedia/analytics/refinery/core/LegacyPageviewDefinition.java
@@ -19,7 +19,7 @@ import java.util.HashSet;
 import java.util.Arrays;
 
 /**
- * Static functions to identify what requests constitute "pageviews",
+ * Functions to identify what requests constitute "pageviews",
  * according to the definition at
  * https://github.com/wikimedia/analytics-refinery/blob/master/oozie/pagecounts-all-sites/load/insert_hourly_pagecounts.hql
  * This is the "legacy" definition, in use by WebStatsCollector and the
@@ -27,35 +27,56 @@ import java.util.Arrays;
  * from 2007 to early 2015, and is to be superseded by the "Pageview" class
  * and isPageview method.
  */
-public class LegacyPageview {
+public class LegacyPageviewDefinition {
 
-    private static final Pattern acceptedUriHostsPattern = Pattern.compile(
+    /*
+     * Meta-methods to enable eager instantiation in a singleton-based way.
+     * in non-Java terms: you get to only create one class instance, and only
+     * when you need it, instead of always having everything (static/eager instantiation)
+     * or always generating everything anew (!singletons). So we have:
+     * (1) an instance;
+     * (2) an empty constructor (to avoid people just calling the constructor);
+     * (3) an actual getInstance method to allow for instantiation.
+     */
+    private static final LegacyPageviewDefinition instance = new LegacyPageviewDefinition();
+
+    private LegacyPageviewDefinition() {
+    }
+
+    public static LegacyPageviewDefinition getInstance(){
+        return instance;
+    }
+
+    /*
+     * Now back to the good part.
+     */
+    private final Pattern acceptedUriHostsPattern = Pattern.compile(
         "\\.(mediawiki|wik(ibooks|idata|imediafoundation|inews|ipedia|iquote|isource|tionary|iversity|ivoyage))\\.org$"
     );
 
-    private static final Pattern acceptedMetaUriHostsPattern = Pattern.compile(
+    private final Pattern acceptedMetaUriHostsPattern = Pattern.compile(
 		    "(commons|incubator|meta|outreach|quality|species|strategy|usability)(\\.m)?\\.wikimedia\\.org$"
     );
 
-    private static final Pattern acceptedUriPattern = Pattern.compile(
+    private final Pattern acceptedUriPattern = Pattern.compile(
         "^/wiki/"
     );
 
-    private static final Pattern rejectedUriPattern = Pattern.compile(
+    private final Pattern rejectedUriPattern = Pattern.compile(
         "^/wiki/Special\\:CentralAutoLogin/"
     );
-    private static final HashSet<String> rejectedUriPathPages = new HashSet<String>(Arrays.asList(
+    private final HashSet<String> rejectedUriPathPages = new HashSet<String>(Arrays.asList(
         "/wiki/undefined",
         "/wiki/Undefined"
     ));
 
-    private static final HashSet<String> rejectedStatusCodes = new HashSet<String>(Arrays.asList(
+    private final HashSet<String> rejectedStatusCodes = new HashSet<String>(Arrays.asList(
         "301",
         "302",
         "303"
     ));
 
-    private static final Pattern rejectedIPPattern = Pattern.compile(
+    private final Pattern rejectedIPPattern = Pattern.compile(
         "^(10\\.20\\.0|10\\.64\\.0|10\\.128\\.0|10\\.64\\.32|208\\.80\\.15[2-5]|91\\.198\\.174)\\..+"
     );
 
@@ -70,7 +91,7 @@ public class LegacyPageview {
      * @param   uriQuery       Query portion of the URI
      * @param   httpStatus     HTTP request status code
      */
-    public static boolean isLegacyPageview(
+    public boolean isLegacyPageview(
         String ip,
         String xForwardedFor,
         String uriHost,
@@ -84,19 +105,19 @@ public class LegacyPageview {
 
             //The host is a "recognised" project
             &&  (
-                    Pageview.patternIsFound(acceptedUriHostsPattern, uriHost)
-                    || Pageview.patternIsFound(acceptedMetaUriHostsPattern, uriHost)
+                    Utilities.patternIsFound(acceptedUriHostsPattern, uriHost)
+                    || Utilities.patternIsFound(acceptedMetaUriHostsPattern, uriHost)
                 )
             //The URI path starts with /wiki/, and
             //isn't to undefined, Undefined or Special:CentralAutoLogin
-            && Pageview.patternIsFound(acceptedUriPattern, uriPath)
-            && !Pageview.patternIsFound(rejectedUriPattern, uriPath)
+            && Utilities.patternIsFound(acceptedUriPattern, uriPath)
+            && !Utilities.patternIsFound(rejectedUriPattern, uriPath)
             && !rejectedUriPathPages.contains(uriPath)
 
             //The source IP isn't in a specified range (or,
             //is, but the XFF field is not empty)
             &&  (
-                    !Pageview.patternIsFound(rejectedIPPattern, ip)
+                    !Utilities.patternIsFound(rejectedIPPattern, ip)
                     || !xForwardedFor.equals("-")
                 )
         );