+++ refinery-hive/src/main/java/org/wikimedia/analytics/refinery/hive/IsCrawlerUDF.java
@@ -27,7 +27,8 @@ public class IsCrawlerUDF extends UDF {
     public boolean evaluate(
         String user_agent
     ) {
-        return Webrequest.isCrawler(
+        Webrequest webrequest_inst = Webrequest.getInstance();
+        return webrequest_inst.isCrawler(
              user_agent
         );
     }