+++ changelog.md
@@ -1,5 +1,6 @@
 ## v0.0.9-SNAPSHOT
-
+* Generic functions used in multiple classes now live in a single "utilities" class.
+* Not everything is static.
 
 ## v0.0.8
 * Stop counting edit attempts as pageviews